% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_SMC.R
\name{ABC_SMC}
\alias{ABC_SMC}
\title{ABC approach to estimate parameters in diversification models.
This function is adapted from Thijs Janzen's NLTT package
 (\url{https://github.com/thijsjanzen/nLTT})}
\usage{
ABC_SMC(
  obs_data,
  sim_function,
  calc_ss_function,
  init_epsilon_values,
  prior_generating_function,
  prior_density_function,
  number_of_particles = 1000,
  sigma = 0.05,
  stop_rate = 0.001,
  num_iterations,
  idparsopt,
  pars,
  ss_set = 1
)
}
\arguments{
\item{obs_data}{A list of simulation output as observation.}

\item{sim_function}{A function to generate simulated data (particles).}

\item{calc_ss_function}{A function to calculate summary statistic distance
between simulated and observed data.}

\item{init_epsilon_values}{A vector of initial epsilon values.}

\item{prior_generating_function}{Function to generate parameters from the
prior distribution.}

\item{prior_density_function}{Function to calculate the prior probability.}

\item{number_of_particles}{The number of particles in each iteration.}

\item{sigma}{Standard deviation of the perturbance distribution.}

\item{stop_rate}{A numeric value which is the boundary to stop the algorithm.}

\item{num_iterations}{The maximum number of iterations.}

\item{idparsopt}{The id of the parameters that need to be inferred, the others
are fixed.}

\item{ss_set}{A numeric indicates which set of summary statistics that
are used to calculate the distance.}
}
\description{
ABC approach to estimate parameters in diversification models.
This function is adapted from Thijs Janzen's NLTT package
 (\url{https://github.com/thijsjanzen/nLTT})
}
\references{
Janzen, T., Höhna, S. and Etienne, R.S. (2015), Approximate
Bayesian Computation of diversification rates from molecular phylogenies:
introducing a new efficient summary statistic, the nLTT. Methods Ecol Evol,
6: 566-575. https://doi.org/10.1111/2041-210X.12350
}
