% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_SMC_bisse.R
\name{ABC_SMC_bisse}
\alias{ABC_SMC_bisse}
\title{Using ABC approach on DAISIE to estimate CES rates.}
\usage{
ABC_SMC_bisse(
  obs_data,
  sim_function,
  calc_ss_function,
  init_epsilon_values,
  prior_generating_function,
  prior_density_function,
  number_of_particles = 1000,
  sigma = 0.05,
  stop_rate = 0.001,
  num_iterations,
  idparsopt,
  fixpars,
  ss_set = 1
)
}
\description{
Using ABC approach on DAISIE to estimate CES rates.
}
