% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_SMC_DAISIE.R
\name{ABC_SMC_DAISIE}
\alias{ABC_SMC_DAISIE}
\title{Using ABC approach on DAISIE to estimate CES rates.}
\usage{
ABC_SMC_DAISIE(
  obs_data,
  sim_function,
  init_epsilon_values,
  prior_generating_function,
  prior_density_function,
  number_of_particles = 1000,
  sigma = 0.05,
  stop_rate = 0.001,
  replicates = 1,
  num_iterations,
  K,
  idparsopt,
  fixpars
)
}
\value{

}
\description{
Using ABC approach on DAISIE to estimate CES rates.
}
\author{

}
