% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_with_ext.R
\name{calc_clade_size_error}
\alias{calc_clade_size_error}
\title{Calculate error of clade size standard deviation between two simulated data.}
\usage{
calc_clade_size_error(sim_1, sim_2)
}
\arguments{
\item{sim_1}{DAISIE simulation.}

\item{sim_2}{DAISIE simulation.}
}
\value{
A numeric for clade size standard  ddeviation.
}
\description{
Calculate error of clade size standard deviation between two simulated data.
}
\author{
Shu Xie
}
