% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error.R
\name{calc_colon_time_error}
\alias{calc_colon_time_error}
\title{Calculate error of colonization time standard deviation between two simulated data.}
\usage{
calc_colon_time_error(sim_1, sim_2)
}
\arguments{
\item{sim_1}{DAISIE simulation.}

\item{sim_2}{DAISIE simulation.}
}
\value{
A numeric for colonization time standard  deviation.
}
\description{
Calculate error of colonization time standard deviation between two simulated data.
}
\author{
Shu Xie
}
