% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musse_calc_error.R
\name{calc_D_drop}
\alias{calc_D_drop}
\title{function to calculate statistic D when drop tips until only two states exist}
\usage{
calc_D_drop(phy, obs_trait)
}
\description{
function to calculate statistic D when drop tips until only two states exist
}
