% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musse_calc_error.R
\name{calc_D_trans}
\alias{calc_D_trans}
\title{function to calculate statistic D when only keep a single state and transfer all
the tips with the other states to "0"}
\usage{
calc_D_trans(sim, keep_state)
}
\description{
function to calculate statistic D when only keep a single state and transfer all
the tips with the other states to "0"
}
