% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_diffs.R
\name{calc_epsilon_init_secsse}
\alias{calc_epsilon_init_secsse}
\title{calculate the initial epsilon}
\usage{
calc_epsilon_init_secsse(sim)
}
\value{

}
\description{
calculate the initial epsilon
}
\author{
Shu Xie
}
