% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error.R
\name{calc_error_all}
\alias{calc_error_all}
\title{Calculate summary statistic distances between two simulated trees
when using all eight statistics}
\usage{
calc_error_all(sim_1, sim_2, replicates = 1, distance_method = "abs")
}
\arguments{
\item{sim_1}{A list of observed data from DAISIE or BiSSE or TraiSIE model.}

\item{sim_2}{A list of simulated data from DAISIE or BiSSE or TraiSIE model.}

\item{replicates}{A numeric for the number of replicates for the
simulations.}

\item{distance_method}{The method of calculating nLTT statistics, choose from
absolute distance or squared distance.}
}
\value{
A vector with five error metrics:
\describe{
  \item{\code{$total_nltt}}{Numeric value of the the distance of nLTT of
  total species.}
  \item{\code{$singleton_nltt}}{Numeric value of the the distance of nLTT of
  singleton-endemic species.}
  \item{\code{$nonend_nltt}}{Numeric value of the the distance of nLTT of
  non-endemic species.}
  \item{\code{$clade_size}}{Numeric value of the the distance of clade size
  standard deviation.}
  \item{\code{$colon_time}}{Numeric value of the the distance of colonization
  time standard deviation.}
  \item{\code{$num_total}}{Numeric value of the the distance of the number of
  total species.}
  \item{\code{$num_end}}{Numeric value of the the distance of the number of
  endemic-endemic species.}
  \item{\code{$num_nonend}}{Numeric value of the the distance of the number of
  non-endemic species.}
}
}
\description{
Calculate summary statistic distances between two simulated trees
when using all eight statistics
}
\author{
Shu Xie
}
