% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisse_calc_error.R
\name{calc_error_bisse}
\alias{calc_error_bisse}
\title{Calculates error metrics between two simulations
NLTTs_D}
\usage{
calc_error_bisse(sim_1, sim_2, distance_method = "abs")
}
\description{
Calculates error metrics between two simulations
NLTTs_D
}
