% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosse_calc_error.R
\name{calc_error_geosse_D}
\alias{calc_error_geosse_D}
\title{Calculates error metrics between two simulations under geosse model
ss0 NLTTs_Delta}
\usage{
calc_error_geosse_D(sim_1, sim_2, distance_method = "abs")
}
\description{
Calculates error metrics between two simulations under geosse model
ss0 NLTTs_Delta
}
