% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosse_calc_error.R
\name{calc_error_geosse_Ddrop_M}
\alias{calc_error_geosse_Ddrop_M}
\title{ss3 Calculates error metrics between two simulations under geosse model
NLTTs_Delta}
\usage{
calc_error_geosse_Ddrop_M(sim_1, sim_2, distance_method = "abs")
}
\description{
ss3 Calculates error metrics between two simulations under geosse model
NLTTs_Delta
}
