% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosse_calc_error.R
\name{calc_error_geosse_Dtrans_M}
\alias{calc_error_geosse_Dtrans_M}
\title{ss4 calculate NLTT statistics + transfered D statistic (e.g. S1 VS Sothers)}
\usage{
calc_error_geosse_Dtrans_M(sim_1, sim_2, distance_method = "abs")
}
\description{
ss4 calculate NLTT statistics + transfered D statistic (e.g. S1 VS Sothers)
}
