% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musse_calc_error.R
\name{calc_error_musse_D_trans}
\alias{calc_error_musse_D_trans}
\title{NLTTs_Phylosig M
calculate NLTT statistics + transfered D statistic (e.g. S1 VS Sothers)}
\usage{
calc_error_musse_D_trans(sim_1, sim_2, distance_method = "abs")
}
\description{
NLTTs_Phylosig M
calculate NLTT statistics + transfered D statistic (e.g. S1 VS Sothers)
}
