% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musse_calc_error.R
\name{calc_error_musse_Delta}
\alias{calc_error_musse_Delta}
\title{Calculates error metrics between two simulations under musse model
NLTTs_Delta}
\usage{
calc_error_musse_Delta(sim_1, sim_2, distance_method = "abs")
}
\description{
Calculates error metrics between two simulations under musse model
NLTTs_Delta
}
