% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error.R
\name{calc_error_nltt}
\alias{calc_error_nltt}
\title{Calculate summary statistic distances between two simulated trees
when using nltt statistics}
\usage{
calc_error_nltt(sim_1, sim_2, replicates, distance_method)
}
\arguments{
\item{sim_1}{A list of observed data from DAISIE or BiSSE or TraiSIE model.}

\item{sim_2}{A list of simulated data from DAISIE or BiSSE or TraiSIE model.}

\item{replicates}{A numeric for the number of replicates for the
simulations.}

\item{distance_method}{The method of calculating nLTT statistics, choose from
absolute distance or squared distance.}
}
\value{
A vector with five error metrics:
\describe{
  \item{\code{$total_nltt}}{Numeric value of the the distance of nLTT of
  total species.}
  \item{\code{$singleton_nltt}}{Numeric value of the the distance of nLTT of
  singleton-endemic species.}
  \item{\code{$nonend_nltt}}{Numeric value of the the distance of nLTT of
  non-endemic species.}
}
}
\description{
Calculate summary statistic distances between two simulated trees
when using nltt statistics
}
\author{
Shu Xie
}
