% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_daisie_pairwise.R
\name{calc_error_pairwise}
\alias{calc_error_pairwise}
\title{This is specific for DAISIE related models}
\usage{
calc_error_pairwise(
  sim_1,
  sim_2,
  replicates,
  distance_method,
  pairwise_method = 1
)
}
\description{
This is specific for DAISIE related models
}
