% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_secsse.R
\name{calc_error_secsse}
\alias{calc_error_secsse}
\title{Calculates error metrics between two simulations}
\usage{
calc_error_secsse(sim_1, sim_2, distance_method = "abs")
}
\description{
Calculates error metrics between two simulations
}
