% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traisie_calc_error.R
\name{calc_error_trait}
\alias{calc_error_trait}
\title{Calculates error metrics between two TraiSIE simulations}
\usage{
calc_error_trait(sim_1, sim_2, replicates, distance_method)
}
\value{
A list with five error metrics:
\describe{
  \item{\code{$spec_nltt_error}}{Numeric vector with the nltt error between
    the two simulations.}
  \item{\code{$num_spec_error}}{Numeric vector with the difference in the
    number of species at the end of the simulation between the two
    simulations.}
  \item{\code{$num_col_error}}{Numeric vector with the difference in the
    number of colonists at the end of the simulation between the two
    simulations.}
  \item{\code{$endemic_nltt_error}}{Numeric vector with the nltt error of the
    endemic species between the two simulations.}
  \item{\code{$nonendemic_nltt_error}}{Numeric vector with the nltt error of
    the non-endemic species between the two simulations.}
}
}
\description{
Calculates error in number of species and colonists, number of
species-through-time, number of endemics-through-time and
non-endemics-through-time.
}
