% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_loglik_prior.R
\name{calc_log_lik_bisse}
\alias{calc_log_lik_bisse}
\title{Calculates the log likelihood of secsse model}
\usage{
calc_log_lik_bisse(params, datalist)
}
\value{
a numeric represents the log likelihood
}
\description{
Calculates the log likelihood of secsse model
}
