% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_loglik_prior.R
\name{calc_log_lik_DAISIE}
\alias{calc_log_lik_DAISIE}
\title{Calculates the posterior probability}
\usage{
calc_log_lik_DAISIE(params, datalist, idparsopt)
}
\value{
a numeric represents the posterior probability
}
\description{
Calculates the posterior probability
}
