% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_loglik_prior.R
\name{calc_log_lik_DAISIE_DI}
\alias{calc_log_lik_DAISIE_DI}
\title{Calculates the posterior probability}
\usage{
calc_log_lik_DAISIE_DI(params, datalist)
}
\value{
a numeric represents the posterior probability
}
\description{
Calculates the posterior probability
}
