% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_secsse.R
\name{calc_log_pp_secsse}
\alias{calc_log_pp_secsse}
\title{Calculates the posterior probability}
\usage{
calc_log_pp_secsse(params, datalist, idparsopt)
}
\value{
a numeric represents the posterior probability
}
\description{
Calculates the posterior probability
}
