% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_loglik_prior.R
\name{calc_log_prior_bisse_logtrans}
\alias{calc_log_prior_bisse_logtrans}
\title{Calculates the log prior density with logtransform}
\usage{
calc_log_prior_bisse_logtrans(params, idparsopt)
}
\value{
a numeric represents the log prior density
}
\description{
Calculates the log prior density with logtransform
}
