% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_loglik_prior.R
\name{calc_log_prior_DAISIE_DI}
\alias{calc_log_prior_DAISIE_DI}
\title{Calculates the log prior density}
\usage{
calc_log_prior_DAISIE_DI(params, idparsopt)
}
\value{
a numeric represents the log prior density
}
\description{
Calculates the log prior density
}
