% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_DAISIE.R
\name{calc_loglik}
\alias{calc_loglik}
\title{Calculates the log likelihood}
\usage{
calc_loglik(params, datalist, idparsopt)
}
\value{
a numeric represents the log likelihood
}
\description{
Calculates the log likelihood
}
