% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_ss.R
\name{calc_num_specs}
\alias{calc_num_specs}
\title{Calculate summary statistic differences between simulated data and a specific
replicate of (simulated) observed data.}
\usage{
calc_num_specs(sim, replicates, distance_method = "abs")
}
\arguments{
\item{sim}{A datalist of observed data with more than one replicate.}

\item{replicates}{The number of replicates used for calculating summary
statistics.}
}
\description{
Calculate summary statistic differences between simulated data and a specific
replicate of (simulated) observed data.
}
\author{
Shu Xie
}
