% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_diffs.R
\name{calc_ss_diff}
\alias{calc_ss_diff}
\title{Calculate summary statistic differences between simulated data and a specific
replicate of (simulated) observed data.}
\usage{
calc_ss_diff(sim1, sim2, ss_set, pairwise_method)
}
\arguments{
\item{sim1}{A datalist of observed data with more than one replicate.}

\item{sim2}{A datalist of simulated data created by DAISIE simulation model.}

\item{replicates}{The number of replicates used for calculating summary
statistics.}
}
\value{
A list with numeric vectors of diff statistics for:
\itemize{
  \item{\code{$spec_nltt_error}}
  \item{\code{$num_spec_error}}
  \item{\code{$num_col_error}}
  \item{\code{$endemic_nltt_error}}
  \item{\code{$nonendemic_nltt_error}}
}
}
\description{
Calculate summary statistic differences between simulated data and a specific
replicate of (simulated) observed data.
}
\author{
Shu Xie
}
