% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_secsse.R
\name{calc_ss_secsse}
\alias{calc_ss_secsse}
\title{Calculate summary statistic differences between simulated data and a specific
replicate of (simulated) observed data.}
\usage{
calc_ss_secsse(sim)
}
\arguments{
\item{sim}{A datalist of observed data with more than one replicate.}

\item{replicates}{The number of replicates used for calculating summary
statistics.}
}
\value{
A list with numeric vectors of diff statistics for:
\itemize{
  \item{\code{$ana_endemic_nltt}}
  \item{\code{$clado_endemic_nltt}}
  \item{\code{$nonendemic_nltt}}
  \item{\code{$num_col_sim}}
  \item{\code{$clade_size_sd}}
  \item{\code{$colon_time_sd}}
}
}
\description{
Calculate summary statistic differences between simulated data and a specific
replicate of (simulated) observed data.
}
\author{
Shu Xie
}
