% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\name{calc_weight}
\alias{calc_weight}
\title{Calculate weight of each particle}
\usage{
calc_weight(
  weights,
  particles,
  current,
  sigma,
  prior_density_function,
  idparsopt
)
}
\arguments{
\item{weights}{A vector of weights}

\item{particles}{A list of parameter combinations}

\item{current}{A vector of current parameter combination to determine the
weight.}

\item{sigma}{Standard deviation of the pertubation.}

\item{prior_density_function}{Function to calculate the prior probability.}
}
\value{
A numeric as the estimated weight
}
\description{
Calculate weight of each particle
}
\author{
Shu Xie
}
