% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error.R
\name{clade_ltt}
\alias{clade_ltt}
\title{Get the number of clades through time from
the colonization times from each exist clade.}
\usage{
clade_ltt(sim, brt)
}
\arguments{
\item{sim}{A list of simulated data from DAISIE or BiSSE or TraiSIE model.}

\item{brt}{A list of vectors containing the branching times in each clade.}
}
\description{
Get the number of clades through time from
the colonization times from each exist clade.
}
\author{
Shu Xie
}
