% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clade_size_error.R
\name{clade_size_sd}
\alias{clade_size_sd}
\title{Calculate clade size standard deviation among clades in a DAISIE simulation.}
\usage{
clade_size_sd(sim)
}
\arguments{
\item{sim}{DAISIE simulation.}
}
\value{
A numeric for clade size standard  ddeviation.
}
\description{
Calculate clade size standard deviation among clades in a DAISIE simulation.
}
\author{
Shu Xie
}
