% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.R
\name{colon_time_sd}
\alias{colon_time_sd}
\title{Calculate colonization time standard deviation among clades in a DAISIE simulation.}
\usage{
colon_time_sd(sim)
}
\arguments{
\item{sim}{DAISIE simulation.}
}
\value{
A numeric for colonization time standard deviation.
}
\description{
Calculate colonization time standard deviation among clades in a DAISIE simulation.
}
\author{
Shu Xie
}
