% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ref_table_geosse.R
\name{create_ref_table_geosse}
\alias{create_ref_table_geosse}
\title{Create reference tables for abcrf}
\usage{
create_ref_table_geosse(save_output = TRUE, id = 1)
}
\arguments{
\item{save_output}{A boolean to determine whether to save or return output.}

\item{id}{A numeric of id of the reference table.}
}
\description{
Create reference tables for abcrf
}
\author{
Shu Xie
}
