% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error.R
\name{full_ltt}
\alias{full_ltt}
\title{Calculate summary statistic distances between two simulated trees
when using phylogenetic statistics}
\usage{
full_ltt(sim, brt)
}
\arguments{
\item{sim}{A list of simulated data from DAISIE or BiSSE or TraiSIE model.}

\item{brt}{A list of vectors containing the branching times in each clade.}
}
\description{
Calculate summary statistic distances between two simulated trees
when using phylogenetic statistics
}
\author{
Shu Xie
}
