% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_bisse_sim}
\alias{get_bisse_sim}
\title{Simulation function to create simulations in ABC.}
\usage{
get_bisse_sim(parameters, pool_init_states, replicates = 1)
}
\arguments{
\item{parameters}{A vector for rates.}

\item{pool_init_states}{Root state pool.}

\item{replicates}{The number of replicates for bisse simulation.}
}
\value{
A list contains simulated trees
}
\description{
Simulation function to create simulations in ABC.
}
