% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_DAISIE_sim_DI}
\alias{get_DAISIE_sim_DI}
\title{Simulation function to create DAISIE simulations}
\usage{
get_DAISIE_sim_DI(parameters, replicates = 1)
}
\arguments{
\item{parameters}{A vector for CES rates.}

\item{replicates}{The number of replicates(islands) for DAISIE simulation.}

\item{K}{Carrying capacity, Inf for diverdity-independent models.}
}
\value{
A list contains simulated islands
}
\description{
Simulation function to create DAISIE simulations
}
