% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_geosse_sim}
\alias{get_geosse_sim}
\title{Simulation function to create simulations in ABC.}
\usage{
get_geosse_sim(parameters, replicates = 1)
}
\arguments{
\item{parameters}{A vector for CES rates.}

\item{replicates}{The number of replicates(islands) for geosse simulation.}
}
\value{
A list contains simulated islands
}
\description{
Simulation function to create simulations in ABC.
}
