% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_geosse_sim_create_obs}
\alias{get_geosse_sim_create_obs}
\title{Simulation function to create GeoSSE simulations as observed data}
\usage{
get_geosse_sim_create_obs(parameters, replicates = 1)
}
\arguments{
\item{parameters}{A vector for CES rates.}

\item{replicates}{The number of replicates for geosse simulation.}
}
\value{
A list contains simulated trees
}
\description{
Simulation function to create GeoSSE simulations as observed data
}
