% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_musse_sim_create_obs}
\alias{get_musse_sim_create_obs}
\title{Simulation function to create MuSSE simulations as observed data}
\usage{
get_musse_sim_create_obs(parameters, pool_init_states, replicates = 1)
}
\arguments{
\item{parameters}{A vector for rates.}

\item{replicates}{The number of replicates for musse simulation.}
}
\value{
A list contains simulated trees
}
\description{
Simulation function to create MuSSE simulations as observed data
}
