% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obs_sim.R
\name{load_obs_sim}
\alias{load_obs_sim}
\title{Loads the parameter space into the environment for simulations}
\usage{
load_obs_sim(scenario)
}
\arguments{
\item{scenario}{A string of a scenario name which contains a series of
parameter sets.}
}
\value{
list of observed data.
}
\description{
Loads the parameter space into the environment for simulations
}
\author{
Shu Xie
}
