% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_param_space.R
\name{load_param_space}
\alias{load_param_space}
\title{Loads the parameter space into the environment for simulations}
\usage{
load_param_space(param_space_name)
}
\value{
parameter space.
}
\description{
Loads the parameter space into the environment for simulations
}
\author{
Shu Xie
}
