% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scenario.R
\name{load_scenario}
\alias{load_scenario}
\title{Loads the parameter sets into the environment for simulations}
\usage{
load_scenario(scenario)
}
\value{
tibble parameter space in a certain scenario.
}
\description{
Loads the parameter sets into the environment for simulations
}
\author{
Shu Xie
}
