% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{mcmc
This function is adapted from Thijs Janzen's NLTT package
 (\url{https://github.com/thijsjanzen/nLTT})}
\usage{
MCMC(
  datalist,
  log_lik_function,
  log_prior_function,
  logform = FALSE,
  parameters,
  iterations,
  burnin,
  thinning = 1,
  sigma = 1,
  idparsopt
)
}
\description{
mcmc
This function is adapted from Thijs Janzen's NLTT package
 (\url{https://github.com/thijsjanzen/nLTT})
}
\references{
Janzen, T., Höhna, S. and Etienne, R.S. (2015), Approximate
Bayesian Computation of diversification rates from molecular phylogenies:
introducing a new efficient summary statistic, the nLTT. Methods Ecol Evol,
6: 566-575. https://doi.org/10.1111/2041-210X.12350
}
