% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error_pairwise.R
\name{pairwise_cross}
\alias{pairwise_cross}
\title{Get the pairwise distance across all the clades}
\usage{
pairwise_cross(sim1, sim2)
}
\description{
Get the pairwise distance across all the clades
}
