% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_calc_error.R
\name{pairwise_sort_per_clade}
\alias{pairwise_sort_per_clade}
\title{pirewise comparison between two simulations among clades(keep the same
number of clades)}
\usage{
pairwise_sort_per_clade(brts1, brts2)
}
\description{
pirewise comparison between two simulations among clades(keep the same
number of clades)
}
