% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_prior_functions.R
\name{prior_dens_DI}
\alias{prior_dens_DI}
\title{prior density function}
\usage{
prior_dens_DI(pars, idparsopt)
}
\value{
Density of the given parameter set
}
\description{
prior density function
}
