% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisse_prior_functions.R
\name{prior_dens_geosse}
\alias{prior_dens_geosse}
\title{prior density function for geosse model}
\usage{
prior_dens_geosse(pars, idparsopt)
}
\value{
Density of the given parameter set
}
\description{
prior density function for geosse model
}
