% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisse_prior_functions.R
\name{prior_dens_musse}
\alias{prior_dens_musse}
\title{prior density function for bisse model}
\usage{
prior_dens_musse(pars, idparsopt)
}
\value{
Density of the given parameter set
}
\description{
prior density function for bisse model
}
