% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traisie_prior_functions.R
\name{prior_dens_trait}
\alias{prior_dens_trait}
\title{prior density function for TraiSIE}
\usage{
prior_dens_trait(pars, idparsopt)
}
\value{
a numeric represents the log likelihood
}
\description{
prior density function for TraiSIE
}
