% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{prior_gen}
\alias{prior_gen}
\title{prior function to generate parameters}
\usage{
prior_gen(idparsopt)
}
\value{
a numeric represents the log likelihood
}
\description{
prior function to generate parameters
}
