% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_prior_functions.R
\name{prior_gen_DI}
\alias{prior_gen_DI}
\title{prior function to generate parameters}
\usage{
prior_gen_DI(pars, idparsopt)
}
\value{
a vector of parameters
}
\description{
prior function to generate parameters
}
