% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisse_prior_functions.R
\name{prior_gen_geosse}
\alias{prior_gen_geosse}
\title{prior function to generate parameters for geosse}
\usage{
prior_gen_geosse(pars, idparsopt)
}
\value{
a vector of parameters
}
\description{
prior function to generate parameters for geosse
}
