% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse_prior_functions.R
\name{prior_gen_musse}
\alias{prior_gen_musse}
\title{prior function to generate parameters for bisse}
\usage{
prior_gen_musse(pars, idparsopt)
}
\value{
a vector of parameters
}
\description{
prior function to generate parameters for bisse
}
