% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions_trait.R
\name{prior_gen_trait}
\alias{prior_gen_trait}
\title{prior function to generate TraiSIE parameters}
\usage{
prior_gen_trait(pars, idparsopt)
}
\value{
a numeric represents the log likelihood
}
\description{
prior function to generate TraiSIE parameters
}
