% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ABC.R
\name{run_ABC}
\alias{run_ABC}
\title{Run ABC}
\usage{
run_ABC(
  param_space_name,
  param_set,
  idparsopt,
  sim_model = "DAISIE",
  save_output = TRUE,
  ss_set = 1,
  pairwise_method
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"DAISIE_ABC"}, \code{"DAISIE_MCMC"},\code{"TraiSIE_ABC"}}

\item{param_set}{A numeric of a specific line of parameter set in parameter space}

\item{idparsopt}{A vector of positions of the parameters that be optimized.}

\item{sim_model}{\code{"DAISIE"}, \code{"TraiSIE"}}

\item{save_output}{A boolean to determine whether to save or return output.}
}
\value{

}
\description{
Run ABC
}
\author{
Shu Xie
}
