% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ABC_trait_DAISIE.R
\name{run_ABC_TraiSIE}
\alias{run_ABC_TraiSIE}
\title{Run ABC_TRAISIE}
\usage{
run_ABC_TraiSIE(param_space_name, param_set, idparsopt, save_output = TRUE)
}
\value{

}
\description{
Run ABC_TRAISIE
}
\author{
Shu Xie
}
