% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ABC_trait_DAISIE.R
\name{run_ABC_trait_DAISIE}
\alias{run_ABC_trait_DAISIE}
\title{Run ABC_TRAISIE}
\usage{
run_ABC_trait_DAISIE(
  param_space_name,
  param_set,
  save_output = TRUE,
  with_trait = FALSE
)
}
\value{

}
\description{
Run ABC_TRAISIE
}
\author{
Shu Xie
}
