% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc_secsse.R
\name{run_MCMC_secsse}
\alias{run_MCMC_secsse}
\title{Run mcmc}
\usage{
run_MCMC_secsse(param_space_name, param_set, idparsopt, save_output = TRUE)
}
\value{

}
\description{
Run mcmc
}
\author{
Shu Xie
}
