% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_diffs.R
\name{select_ss_multi}
\alias{select_ss_multi}
\title{initial epsilon based on selected summary statistic}
\usage{
select_ss_multi(ss_set)
}
\value{

}
\description{
initial epsilon based on selected summary statistic
}
\author{
Shu Xie
}
