% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_diffs.R
\name{select_ss_secsse}
\alias{select_ss_secsse}
\title{Select the combination of summary statitsics}
\usage{
select_ss_secsse(ss, ss_set)
}
\arguments{
\item{ss}{A vector contains all the calculated summary statistics}

\item{ss_set}{A numeric to choose which combination of summary statistics}
}
\description{
Select the combination of summary statitsics
}
\author{
Shu Xie
}
