% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_no_ext.R
\name{clade_ltt}
\alias{clade_ltt}
\title{Get the number of clades through time from
the colonization times from each exist clade.}
\usage{
clade_ltt(sim)
}
\arguments{
\item{sim}{A datalist of observed data with more than one replicate.}
}
\description{
Get the number of clades through time from
the colonization times from each exist clade.
}
\author{
Shu Xie
}
