% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(sim1, sim2, replicates, obs_rep)
}
\arguments{
\item{sim1}{A datalist which used to create observed data.}

\item{sim2}{A datalist simulated using DAISIE simulation model.}

\item{replicates}{A numeric for the number of replicates used for calculating summary
statistics.}

\item{obs_rep}{A numeric which means a specific replicate of observed data
used to calculate difference of summary statistics.}
}
\value{
Nothing
}
\description{
Default parameter documentation
}
\author{
Shu Xie
}
\keyword{internal}
