% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_no_ext.R
\name{full_ltt}
\alias{full_ltt}
\title{Get the NLTT dataframe from the branching times from each exist clade.}
\usage{
full_ltt(sim, brt)
}
\arguments{
\item{sim}{A datalist of observed data with more than one replicate.}
}
\description{
Get the NLTT dataframe from the branching times from each exist clade.
}
\author{
Shu Xie
}
