% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_musse_sim}
\alias{get_musse_sim}
\title{Simulation fucntion to create simulations in ABC.}
\usage{
get_musse_sim(parameters, pool_init_states, replicates = 1)
}
\arguments{
\item{parameters}{A vector for CES rates.}

\item{replicates}{The number of replicates(islands) for secsse simulation.}
}
\value{
A list contains simulated islands
}
\description{
Simulation fucntion to create simulations in ABC.
}
\author{
Shu Xie
}
