% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_musse_sim_create_obs}
\alias{get_musse_sim_create_obs}
\title{Simulation fucntion to create MuSSE simualtions as observed data}
\usage{
get_musse_sim_create_obs(parameters, pool_init_states = NA, replicates = 1)
}
\arguments{
\item{parameters}{A vector for CES rates.}

\item{replicates}{The number of replicates(islands) for secsse simulation.}
}
\value{
A list contains simulated islands
}
\description{
Simulation fucntion to create MuSSE simualtions as observed data
}
\author{
Shu Xie
}
