% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obs_sim.R
\name{load_obs_sim}
\alias{load_obs_sim}
\title{Loads the parameter space into the environment for simulations}
\usage{
load_obs_sim(param_space_name)
}
\value{
list of observed data.
}
\description{
Loads the parameter space into the environment for simulations
}
\author{
Shu Xie
}
