% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{mcmc}
\usage{
MCMC(
  datalist,
  log_lik_function,
  log_prior_function,
  parameters,
  iterations,
  burnin = round(iterations/3),
  thinning = 1,
  sigma = 1,
  idparsopt
)
}
\value{

}
\description{
mcmc
}
\author{
Shu Xie
}
