% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_DAISIE.R
\name{MCMC_DAISIE}
\alias{MCMC_DAISIE}
\title{mcmc}
\usage{
MCMC_DAISIE(
  datalist,
  likelihood_function,
  parameters,
  logtransforms = c(TRUE, TRUE, TRUE, TRUE),
  iterations,
  burnin = round(iterations/3),
  thinning = 1,
  sigma = 1,
  idparsopt = 1:4
)
}
\value{

}
\description{
mcmc
}
\author{
Shu Xie
}
