% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions_secsse.R
\name{prior_dens_secsse}
\alias{prior_dens_secsse}
\title{prior density function}
\usage{
prior_dens_secsse(pars, idparsopt)
}
\value{
Density of the given parameter set
}
\description{
prior density function
}
